requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.1.0
audioread==3.0.1
babel==2.17.0
braceexpand==0.1.7
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
cytoolz==1.0.1
darkdetect==0.8.0
datasets==3.3.2
decorator==5.2.1
dill==0.3.8
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
grpcio==1.70.0
gTTS==2.5.4
huggingface-hub==0.29.1
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
intervaltree==3.1.0
ipython==9.0.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.5
jiwer==3.1.0
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
lazy_loader==0.4
lhotse==1.29.0
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.12.0
lilcom==1.8.0
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nemo-toolkit==2.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
optuna==4.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==19.0.1
PyAudio==0.2.14
pycparser==2.22
pygame==2.6.1
Pygments==2.19.1
pygtrie==2.5.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pyttsx3==2.98
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.38
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tkcalendar==1.6.1
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.6.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
triton==3.2.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
webdataset==0.2.111
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
